WIP: Updated docs and moved dataframe accessors into their own files
