+++ clarite/__init__.py
@@ -1,7 +1,7 @@
 # flake8: noqa
 from ._version import get_versions
 
-from .df_accessor import ClariteProcessDFAccessor, ClariteModifyDFAccessor, ClaritePlotDFAccessor, ClariteDescribeDFAccessor
+from .modules.df_accessors import *
 from .modules import process, modify, plot, describe, analyze, io, survey
 
 __version__ = get_versions()['version']
