+++ clarite/modules/describe.py
@@ -1,3 +1,14 @@
+"""
+Describe
+========
+
+Functions that are used to gather information about some data
+
+  **DataFrame Accessor**: ``clarite_describe``
+
+  **CLI Command**: ``describe``
+"""
+
 # Describe - functions that are used to gather information about some data
 
 import numpy as np
