+++ clarite/modules/df_accessors/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+from.clarite_describe import ClariteDescribeDFAccessor
+from .clarite_modify import ClariteModifyDFAccessor
+from .clarite_plot import ClaritePlotDFAccessor
+from .clarite_process import ClariteProcessDFAccessor
