+++ clarite/modules/survey/__init__.py
@@ -1,3 +1,14 @@
 # flake8: noqa
+"""
+Survey
+========
+
+Complex survey design
+
+  **DataFrame Accessor**: ``None``
+
+  **CLI Command**: ``None``
+"""
+
 from .survey_design import SurveyDesignSpec
 from .survey_model import SurveyModel
